[werft] Extract slack notifiation
