+++ .werft/build.js
@@ -4,7 +4,7 @@ const { werft, exec, gitTag } = require('./util/shell.js');
 const { sleep } = require('./util/util.js');
 const { wipeAndRecreateNamespace, setKubectlContextNamespace, deleteNonNamespaceObjects } = require('./util/kubectl.js');
 const { issueAndInstallCertficate } = require('./util/certs.js');
-const https = require('https');
+const { reportBuildFailureInSlack } = require('./util/slack.js');
 
 const GCLOUD_SERVICE_ACCOUNT_PATH = "/mnt/secrets/gcp-sa/service-account.json";
 
@@ -14,49 +14,7 @@ const version = parseVersion(context);
 build(context, version)
     .catch((err) => {
         if (context.Repository.ref === "refs/heads/main") {
-            const repo = context.Repository.host + "/" + context.Repository.owner + "/" + context.Repository.repo;
-            const data = JSON.stringify({
-                "blocks": [
-                    {
-                        "type": "section",
-                        "text": {
-                            "type": "mrkdwn",
-                            "text": ":X: *build failure*\n_Repo:_ `" + repo + "`\n_Build:_ `" + context.Name + "`"
-                        },
-                        "accessory": {
-                            "type": "button",
-                            "text": {
-                                "type": "plain_text",
-                                "text": "Go to Werft",
-                                "emoji": true
-                            },
-                            "value": "click_me_123",
-                            "url": "https://werft.gitpod-dev.com/job/" + context.Name,
-                            "action_id": "button-action"
-                        }
-                    },
-                    {
-                        "type": "section",
-                        "text": {
-                            "type": "mrkdwn",
-                            "text": "```\n" + err + "\n```"
-                        }
-                    }
-                ]
-            });
-            const req = https.request({
-                hostname: "hooks.slack.com",
-                port: 443,
-                path: process.env.SLACK_NOTIFICATION_PATH.trim(),
-                method: "POST",
-                headers: {
-                    'Content-Type': 'application/json',
-                    'Content-Length': data.length,
-                }
-            }, () => process.exit(1));
-            req.on('error', () => process.exit(1));
-            req.write(data);
-            req.end();
+            reportBuildFailureInSlack(context, () => process.exit(1))
         } else {
             process.exit(1);
         }
