+++ .werft/util/slack.js
@@ -1,0 +1,51 @@
+const https = require('https');
+
+function reportBuildFailureInSlack(context, onErr) {
+    const repo = context.Repository.host + "/" + context.Repository.owner + "/" + context.Repository.repo;
+    const data = JSON.stringify({
+        "blocks": [
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": ":X: *build failure*\n_Repo:_ `" + repo + "`\n_Build:_ `" + context.Name + "`"
+                },
+                "accessory": {
+                    "type": "button",
+                    "text": {
+                        "type": "plain_text",
+                        "text": "Go to Werft",
+                        "emoji": true
+                    },
+                    "value": "click_me_123",
+                    "url": "https://werft.gitpod-dev.com/job/" + context.Name,
+                    "action_id": "button-action"
+                }
+            },
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": "```\n" + err + "\n```"
+                }
+            }
+        ]
+    });
+    const req = https.request({
+        hostname: "hooks.slack.com",
+        port: 443,
+        path: process.env.SLACK_NOTIFICATION_PATH.trim(),
+        method: "POST",
+        headers: {
+            'Content-Type': 'application/json',
+            'Content-Length': data.length,
+        }
+    }, onErr);
+    req.on('error', () => process.exit(1));
+    req.write(data);
+    req.end();
+}
+
+module.exports = {
+    reportBuildFailureInSlack: reportBuildFailureInSlack,
+}
